Modify directory structure and update README.md
I moved the relations.py and latticeGenerator.py files to the main
source directory in order to make them more prominent. I presume this
increases the chances that others will find it. I now keep the
unfinished work in the src/ directory.

I also updated the README.md file to explain what the point of the
repository is. It is still very incomplete and should have example usage
plus probably pictures explaining what this program is actually doing.